feat(imgUpload): 分离通用方法
