+++ src/components/base/upload-imgs/index.vue
@@ -3,11 +3,10 @@
  Describe: 多图片上传组件, 附有预览, 排序, 验证等功能
 
 todo: 支持 before-upload
-todo: 支持动态图检测
 todo: 图像验证支持验证是否是动图
 todo: 文档编写
-todo: accept 严格性优化
 todo: jsDoc 编写
+todo: 文件判断使用 serveWorker 优化性能
 -->
 
 <template>
@@ -91,6 +90,13 @@ todo: jsDoc 编写
 </template>
 
 <script>
+import {
+  getFileType,
+  checkIsAnimated,
+  isEmptyObj,
+  createId,
+} from './utils'
+
 /**
  * @typedef {Object<string, number, any>} LocalFileInfo 本地图像通过验证后构造的信息对象
  * @property {string} localSrc 本地图像预览地址
@@ -117,53 +123,6 @@ todo: jsDoc 编写
 
 const ONE_KB = 1024
 const ONE_MB = ONE_KB * 1024
-
-// 检测官方文档: https://mimesniff.spec.whatwg.org/#matching-an-image-type-pattern
-/** 类型检测掩码集合 */
-const patternMask = [{
-  name: 'image/x-icon',
-  mask: [0xFF, 0xFF, 0xFF, 0xFF],
-  byte: [0x00, 0x00, 0x01, 0x00],
-}, {
-  name: 'image/x-icon',
-  mask: [0xFF, 0xFF, 0xFF, 0xFF],
-  byte: [0x00, 0x00, 0x02, 0x00],
-}, {
-  name: 'image/bmp',
-  mask: [0xFF, 0xFF],
-  byte: [0x42, 0x4D],
-}, {
-  name: 'image/gif',
-  mask: [0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF],
-  byte: [0x47, 0x49, 0x46, 0x38, 0x37, 0x61],
-}, {
-  name: 'image/gif',
-  mask: [0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF],
-  byte: [0x47, 0x49, 0x46, 0x38, 0x39, 0x61],
-}, {
-  name: 'image/webp',
-  mask: [0xFF, 0xFF, 0xFF, 0xFF, 0x00, 0x00, 0x00, 0x00, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF],
-  byte: [0x52, 0x49, 0x46, 0x46, 0x00, 0x00, 0x00, 0x00, 0x57, 0x45, 0x42, 0x50, 0x56, 0x50],
-}, {
-  name: 'image/png',
-  mask: [0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF, 0xFF],
-  byte: [0x89, 0x50, 0x4E, 0x47, 0x0D, 0x0A, 0x1A, 0x0A],
-}, {
-  name: 'image/jpeg',
-  mask: [0xFF, 0xFF, 0xFF],
-  byte: [0xFF, 0xD8, 0xFF],
-}]
-
-/** 判断是否是空对象 */
-function isEmptyObj(data) {
-  if (!data) return true
-  return (JSON.stringify(data) === '{}')
-}
-
-/** 生成随机字符串 */
-function createId() {
-  return Math.random().toString(36).substring(2)
-}
 
 /**
  * 创建项, 如不传入参数则创建空项
@@ -244,106 +203,6 @@ function getRangeTip(prx, min, max, unit = '') {
     str += '无限制'
   }
   return str
-}
-
-/**
- * 检测是否是动图
- * 主要针对 Gif 和 Webp 两种格式
- * @param {File} file 需要检测的文件
- * @param {String} fileUrl 文件url
- */
-async function checkIsAnimated({ file, fileUrl, fileType }) {
-  // 参数验证
-  if (!file || !(file instanceof File)) {
-    console.error('isAnimated param check fail: param expected to be File object')
-    return false
-  }
-  // 如果不是 gif 和 webp, 默认作为非动图
-  if (fileType !== 'image/webp' && fileType !== 'image/gif') {
-    return false
-  }
-
-  if (fileType === 'image/webp') {
-    return new Promise((resolve) => {
-      const request = new XMLHttpRequest()
-      request.open('GET', fileUrl, true)
-      request.addEventListener('load', () => {
-        resolve((request.response.indexOf('ANMF') !== -1))
-      })
-      request.send()
-    })
-  }
-  if (fileType === 'image/gif') {
-    return new Promise((resolve) => {
-      const request = new XMLHttpRequest()
-      request.open('GET', fileUrl, true)
-      request.responseType = 'arraybuffer'
-      request.addEventListener('load', () => {
-        const arr = new Uint8Array(request.response)
-        // make sure it's a gif (GIF8)
-        if (arr[0] !== 0x47 || arr[1] !== 0x49 || arr[2] !== 0x46 || arr[3] !== 0x38) {
-          resolve(false)
-          return
-        }
-
-        // ported from php http://www.php.net/manual/en/function.imagecreatefromgif.php#104473
-        // an animated gif contains multiple "frames", with each frame having a
-        // header made up of:
-        // * a static 4-byte sequence (\x00\x21\xF9\x04)
-        // * 4 variable bytes
-        // * a static 2-byte sequence (\x00\x2C) (some variants may use \x00\x21 ?)
-        // We read through the file til we reach the end of the file, or we've found
-        // at least 2 frame headers
-        let frames = 0
-        for (let i = 0, len = arr.length - 9; i < len && frames < 2; ++i) {
-          if (arr[i] === 0x00 && arr[i + 1] === 0x21 && arr[i + 2] === 0xF9 && arr[i + 3] === 0x04 && arr[i + 8] === 0x00 && (arr[i + 9] === 0x2C || arr[i + 9] === 0x21)) {
-            frames++
-          }
-        }
-
-        // if frame count > 1, it's animated
-        resolve(frames > 1)
-      })
-      request.send()
-    })
-  }
-}
-
-/**
- * 检测文件类型
- * 使用文件编码进行检测
- * 支持模式参看: patternMask 定义
- */
-async function getFileType(file) {
-  if (!(file instanceof File)) {
-    return 'unknown'
-  }
-  return new Promise((resolve) => {
-    const fileReader = new FileReader()
-    fileReader.onloadend = (e) => {
-      const header = (new Uint8Array(e.target.result)).slice(0, 20)
-      let type = 'unknown'
-
-      // eslint-disable-next-line arrow-body-style
-      const index = patternMask.findIndex((item) => {
-        // eslint-disable-next-line arrow-body-style
-        return item.mask.every((subItem, subI) => {
-          // subItem 掩码标志
-          // item.byte[subI] 规范值
-          // header[subI] 文件实际值
-          // eslint-disable-next-line
-          return ((subItem & (header[subI] ^ item.byte[subI])) === 0)
-        })
-      })
-
-      if (index >= 0) {
-        type = patternMask[index].name
-      }
-
-      resolve(type)
-    }
-    fileReader.readAsArrayBuffer(file)
-  })
 }
 
 /** for originUpload: 一次请求最多的文件数量 */
